add skipped test for the API, until Next.js lands better API route testing
