+++ .vn/tests/vn/mongoConnection.server.test.ts
@@ -5,21 +5,25 @@
 import {
   connectToDb,
   closeDbConnection,
-} from "../../../src/api/middlewares/mongoConnection";
+} from "../../../src/api/mongoose/connection";
 import mongoose from "mongoose";
 
+if (!process.env.MONGO_URI)
+  throw new Error("MONGO_URI env variable not defined");
+
+const mongoUri = process.env.MONGO_URI;
 describe("api/middlewares/mongoConnection", () => {
   afterEach(async () => {
     await closeDbConnection();
   });
   it("connects to mongo db", async () => {
-    await connectToDb(process.env.MONGO_URI).then(() => {
+    await connectToDb(mongoUri).then(() => {
       expect(mongoose.connection.readyState).toEqual(1);
     });
   });
   it("connects only once if already connecting", async () => {
-    const promise = connectToDb(process.env.MONGO_URI); // you can define a .env.test to configure this
-    const newPromise = connectToDb(process.env.MONGO_URI); // you can define a .env.test to configure this
+    const promise = connectToDb(mongoUri); // you can define a .env.test to configure this
+    const newPromise = connectToDb(mongoUri); // you can define a .env.test to configure this
     expect(promise).toEqual(newPromise);
   });
 });
