+++ .vn/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../src/tsconfig.json",
+  "include": ["./**/*.tsx", "./**/*.ts"],
+}
