+++ package.json
@@ -62,9 +62,10 @@
     "@mdx-js/react": "^1.6.13",
     "@next/mdx": "^10.0.2",
     "@vulcanjs/demo": "^0.0.7",
+    "@vulcanjs/graphql": "^0.2.4",
     "@vulcanjs/mdx": "^0.2.3-alpha.0",
     "@vulcanjs/meteor-legacy": "^0.1.8",
-    "@vulcanjs/mongo": "^0.1.8",
+    "@vulcanjs/mongo": "^0.2.4",
     "@vulcanjs/react-hooks": "^0.2.3",
     "@vulcanjs/react-ui": "^0.2.3",
     "apollo-server-express": "2.14.2",
@@ -115,7 +116,6 @@
     "react-is": "^16.13.1",
     "source-map-support": "^0.5.19",
     "storybook-css-modules-preset": "^1.1.1",
-    "supertest": "^6.1.6",
     "ts-loader": "^7.0.5",
     "ts-node": "^8.10.2",
     "typescript": "=4.3.5",
